Перенес ItemTouchHelper в отдельный класс
