+++ app/src/main/java/com/example/materialdesign/MyItemTouchHelper.kt
@@ -1,0 +1,47 @@
+package com.example.materialdesign
+
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.RecyclerView
+import com.example.materialdesign.interfaces.DraggableViewHolder
+
+class MyItemTouchHelper(adapter: CardRecyclerAdapter) {
+
+    val itemTouchHelper = ItemTouchHelper(object : ItemTouchHelper.Callback() {
+        override fun getMovementFlags(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder): Int {
+            val dragFlags = ItemTouchHelper.UP or ItemTouchHelper.DOWN
+            val swipeFlags = ItemTouchHelper.START or ItemTouchHelper.END
+            return makeMovementFlags(dragFlags, swipeFlags)
+        }
+
+        override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder, target: RecyclerView.ViewHolder): Boolean {
+            adapter.move(viewHolder.adapterPosition, target.adapterPosition)
+            return true
+        }
+
+        override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+            adapter.delete(viewHolder.adapterPosition)
+        }
+
+        override fun onSelectedChanged(viewHolder: RecyclerView.ViewHolder?, actionState: Int) {
+            super.onSelectedChanged(viewHolder, actionState)
+            if (actionState == ItemTouchHelper.ACTION_STATE_IDLE) {
+                return
+            }
+            if (viewHolder is DraggableViewHolder) {
+                viewHolder.onSelected()
+            }
+        }
+
+        override fun clearView(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder) {
+            super.clearView(recyclerView, viewHolder)
+            if (viewHolder is DraggableViewHolder) {
+                viewHolder.onClear()
+            }
+        }
+    })
+
+    fun attachToRecyclerView(recyclerView: RecyclerView) {
+        itemTouchHelper.attachToRecyclerView(recyclerView)
+    }
+
+}
