+++ app/src/main/java/com/example/materialdesign/interfaces/DraggableViewHolder.kt
@@ -1,0 +1,6 @@
+package com.example.materialdesign.interfaces
+
+interface DraggableViewHolder {
+    fun onSelected()
+    fun onClear()
+}
